// src/theme/types.ts
export type ThemeChoice = 'system' | 'light' | 'dark';

export type ThemeColors = {
  bg: string;
  bg2: string;
  ui: string;
  ui2: string;
  tx: string;
  tx2: string;
  tx3: string;
  accent: string;
  accent2: string;
  red: string;
  yellow: string;
  overlay: string;
  status: 'light' | 'dark';
};